added module option for changing dchannel (e1)
when registering devices, the name is filled with device's name, number and port.

options in layer2.c were investigated wong.

tested PRI NT-mode
-> it works, also change mode
tested PRI TE-mode
-> it works, also change mode

	modified:   drivers/isdn/hardware/mISDN/hfc_multi.c
	modified:   drivers/isdn/hardware/mISDN/hfc_pci.c
	modified:   drivers/isdn/hardware/mISDN/hfcs_usb.c
	modified:   drivers/isdn/mISDN/core.c
	modified:   drivers/isdn/mISDN/l1oip_core.c
	modified:   drivers/isdn/mISDN/layer2.c
	modified:   drivers/isdn/mISDN/layer2.h
	modified:   drivers/isdn/mISDN/stack.c
	modified:   include/linux/mISDNif.h

+++ drivers/isdn/mISDN/layer2.h
@@ -81,7 +81,7 @@ enum {
 
 #define L2_STATE_COUNT (ST_L2_8+1)
 
-extern struct layer2	*create_l2(struct mISDNchannel *, u_int, u_int, u_long arg);
+extern struct layer2	*create_l2(struct mISDNchannel *, u_int, u_long, u_long arg);
 extern int		tei_l2(struct layer2 *, u_int, u_long arg);
